WW-5284 Refactor ActionValidatorManager implementations
